LPD-90910 Gate per-parent friendly URL behind LPD-70396

+++ modules/apps/asset/asset-categories-service/src/main/java/com/liferay/asset/categories/internal/model/listener/AssetCategoryFriendlyURLModelListener.java
@@ -11,6 +11,7 @@ import com.liferay.exportimport.kernel.lar.ExportImportThreadLocal;
 import com.liferay.friendly.url.service.FriendlyURLEntryLocalService;
 import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.model.BaseModelListener;
 import com.liferay.portal.kernel.model.ModelListener;
 import com.liferay.portal.kernel.service.ClassNameLocalService;
@@ -29,6 +30,12 @@ public class AssetCategoryFriendlyURLModelListener
 	@Override
 	public void onAfterCreate(AssetCategory assetCategory)
 		throws ModelListenerException {
+
+		if (!FeatureFlagManagerUtil.isEnabled(
+				assetCategory.getCompanyId(), "LPD-70396")) {
+
+			return;
+		}
 
 		try {
 			if (ExportImportThreadLocal.isImportInProcess() ||
@@ -58,6 +65,12 @@ public class AssetCategoryFriendlyURLModelListener
 	@Override
 	public void onBeforeRemove(AssetCategory assetCategory)
 		throws ModelListenerException {
+
+		if (!FeatureFlagManagerUtil.isEnabled(
+				assetCategory.getCompanyId(), "LPD-70396")) {
+
+			return;
+		}
 
 		_friendlyURLEntryLocalService.deleteFriendlyURLEntry(
 			assetCategory.getGroupId(),
